# apps/api/src/workspaces/api-tokens.controller.ts
import {
  Controller,
  Get,
  Post,
  Delete,
  Body,
  Param,
  UseGuards,
  BadRequestException,
  NotFoundException,
  ForbiddenException,
} from '@nestjs/common';
import { ApiTags, ApiOperation, ApiResponse, ApiBearerAuth, ApiParam, ApiBody, ApiProperty } from '@nestjs/swagger';
import { AuthGuard } from '../auth/auth.guard';
import { CurrentUser } from '../auth/current-user.decorator';
import { prisma } from '@repo/database';
import type { User } from '@repo/database';
import { z } from 'zod';
import * as crypto from 'crypto';
import { IsString, IsOptional, IsNumber, IsObject } from 'class-validator';

class CreateApiTokenDto {
  @IsString()
  @ApiProperty({ example: 'My API Token' })
  name: string;

  @IsObject()
  @ApiProperty({
    type: 'object',
    properties: {
      actions: {
        type: 'array',
        items: {
          type: 'string',
          enum: [
            'read:members',
            'write:members',
            'read:departments',
            'write:departments',
            'read:teams',
            'write:teams',
            'read:announcements',
            'write:announcements',
            'read:channels',
            'write:channels',
            'send:messages',
          ],
        },
      },
      departments: { type: 'array', items: { type: 'string' } },
      teams: { type: 'array', items: { type: 'string' } },
    },
  })
  permissions: {
    actions: string[];
    departments?: string[];
    teams?: string[];
  };

  @IsNumber()
  @IsOptional()
  @ApiProperty({ required: false, default: 1000, example: 5000 })
  rateLimit?: number;

  @IsString()
  @IsOptional()
  @ApiProperty({ required: false, description: 'ISO format datetime' })
  expiresAt?: string;
}

const createTokenSchema = z.object({
  name: z.string().min(1).max(100),
  permissions: z.object({
    departments: z.array(z.string()).optional(),
    teams: z.array(z.string()).optional(),
    actions: z.array(
      z.enum([
        'read:members',
        'write:members',
        'read:departments',
        'write:departments',
        'read:teams',
        'write:teams',
        'read:announcements',
        'write:announcements',
        'read:channels',
        'write:channels',
        'send:messages',
      ])
    ),
  }),
  rateLimit: z.number().min(100).max(100000).optional().default(1000),
  expiresAt: z.string().datetime().optional(),
});

@ApiTags('API Tokens')
@ApiBearerAuth()
@Controller('workspaces/:slug/api-tokens')
@UseGuards(AuthGuard)
export class ApiTokensController {
  @Get()
  @ApiOperation({ summary: 'Get all API tokens for a workspace' })
  @ApiParam({ name: 'slug', description: 'The workspace slug' })
  @ApiResponse({ status: 200, description: 'List of API tokens' })
  async getApiTokens(@CurrentUser() user: User, @Param('slug') slug: string) {
    /**
     * ⚡ Performance Optimization:
     * 1. Consolidates workspace lookup, membership verification, and API token retrieval into a single query.
     * 2. Uses nested 'select' to fetch only required fields and relations.
     * 3. Reduces database round-trips from 3 down to 1.
     */
    const workspace = await prisma.workspace.findUnique({
      where: { slug },
      select: {
        id: true,
        members: {
          where: { userId: user.id },
          select: { role: true },
        },
        apiTokens: {
          select: {
            id: true,
            name: true,
            token: true,
            permissions: true,
            rateLimit: true,
            expiresAt: true,
            lastUsedAt: true,
            usageCount: true,
            createdAt: true,
            createdBy: { select: { id: true, name: true, avatar: true } },
          },
          orderBy: { createdAt: 'desc' },
        },
      },
    });

    if (!workspace) {
      throw new NotFoundException('Workspace not found');
    }

    const member = workspace.members[0];

    if (!member || !['owner', 'admin'].includes(member.role)) {
      throw new ForbiddenException('Forbidden');
    }

    const tokens = workspace.apiTokens;

    const maskedTokens = tokens.map(t => ({
      ...t,
      token: `wst_${'*'.repeat(24)}${t.token.slice(-8)}`,
    }));

    return { tokens: maskedTokens };
  }

  @Post()
  @ApiOperation({ summary: 'Create a new API token for a workspace' })
  @ApiParam({ name: 'slug', description: 'The workspace slug' })
  @ApiBody({ type: CreateApiTokenDto })
  @ApiResponse({ status: 201, description: 'API token created' })
  async createApiToken(@CurrentUser() user: User, @Param('slug') slug: string, @Body() body: CreateApiTokenDto) {
    const workspace = await prisma.workspace.findUnique({
      where: { slug },
      select: {
        id: true,
        members: {
          where: { userId: user.id },
          select: { role: true },
        },
      },
    });

    if (!workspace) {
      throw new NotFoundException('Workspace not found');
    }

    const member = workspace.members[0];

    if (!member || !['owner', 'admin'].includes(member.role)) {
      throw new ForbiddenException('Forbidden');
    }

    const validatedData = createTokenSchema.safeParse(body);
    if (!validatedData.success) {
      throw new BadRequestException(validatedData.error.issues);
    }
    const data = validatedData.data;

    const token = `wst_${crypto.randomBytes(32).toString('hex')}`;

    const apiToken = await prisma.workspaceApiToken.create({
      data: {
        workspaceId: workspace.id,
        name: data.name,
        token,
        permissions: data.permissions as any,
        rateLimit: data.rateLimit,
        expiresAt: data.expiresAt ? new Date(data.expiresAt) : null,
        createdById: user.id,
      },
    });

    /**
     * ⚡ Performance Optimization:
     * Background non-critical audit log creation with `.catch()` to avoid blocking response path.
     */
    prisma.workspaceAuditLog
      .create({
        data: {
          workspaceId: workspace.id,
          userId: user.id,
          action: 'api_token.created',
          resource: 'api_token',
          resourceId: apiToken.id,
          metadata: { name: data.name, permissions: data.permissions.actions },
        },
      })
      .catch(() => {});

    return {
      ...apiToken,
      token,
    };
  }

  @Delete(':tokenId')
  @ApiOperation({ summary: 'Delete an API token' })
  @ApiParam({ name: 'slug', description: 'The workspace slug' })
  @ApiParam({ name: 'tokenId', description: 'The token ID' })
  @ApiResponse({ status: 200, description: 'API token deleted' })
  async deleteApiToken(@CurrentUser() user: User, @Param('slug') slug: string, @Param('tokenId') tokenId: string) {
    /**
     * ⚡ Performance Optimization:
     * 1. Uses atomic `deleteMany` with compound workspaceId filter to verify workspace boundary & delete in 1 operation.
     * 2. Backgrounds non-critical audit log creation with `.catch()` to eliminate secondary DB write latency.
     * Expected impact: Eliminates separate read/delete sequences and background write latency.
     */
    const workspace = await prisma.workspace.findUnique({
      where: { slug },
      select: {
        id: true,
        members: {
          where: { userId: user.id },
          select: { role: true },
        },
      },
    });

    if (!workspace) {
      throw new NotFoundException('Workspace not found');
    }

    const member = workspace.members[0];

    if (!member || !['owner', 'admin'].includes(member.role)) {
      throw new ForbiddenException('Forbidden');
    }

    const result = await prisma.workspaceApiToken.deleteMany({
      where: { id: tokenId, workspaceId: workspace.id },
    });

    if (result.count === 0) {
      throw new NotFoundException('API token not found');
    }

    prisma.workspaceAuditLog
      .create({
        data: {
          workspaceId: workspace.id,
          userId: user.id,
          action: 'api_token.deleted',
          resource: 'api_token',
          resourceId: tokenId,
          metadata: {},
        },
      })
      .catch(() => {});

    return { message: 'API token deleted successfully' };
  }
}