# apps/docs/src/content/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v3/workspaces/{slug}": {
      "get": {
        "description": "Retrieve details of a specific workspace by its slug.",
        "operationId": "V3WorkspacesController_getWorkspaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "workspace": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string",
                              "nullable": true
                            },
                            "icon": {
                              "type": "string",
                              "nullable": true
                            },
                            "industry": {
                              "type": "string",
                              "nullable": true
                            },
                            "brandingConfig": {
                              "type": "object",
                              "nullable": true
                            },
                            "createdAt": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Missing provisioning:workspaces scope or unauthorized workspace access."
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get workspace details (Enterprise M2M)",
        "tags": [
          "V3 Workspaces"
        ]
      },
      "patch": {
        "description": "Update the configuration and metadata of a specific workspace.",
        "operationId": "V3WorkspacesController_updateWorkspace",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V3UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "workspace": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            },
                            "description": {
                              "type": "string",
                              "nullable": true
                            },
                            "icon": {
                              "type": "string",
                              "nullable": true
                            },
                            "industry": {
                              "type": "string",
                              "nullable": true
                            },
                            "brandingConfig": {
                              "type": "object",
                              "nullable": true
                            },
                            "updatedAt": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Missing provisioning:workspaces scope or unauthorized access."
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a workspace (Enterprise M2M)",
        "tags": [
          "V3 Workspaces"
        ]
      },
      "delete": {
        "description": "Permanently deletes a specific workspace.",
        "operationId": "V3WorkspacesController_deleteWorkspace",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden: Missing provisioning:workspaces scope or unauthorized access."
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a workspace (Enterprise M2M)",
        "tags": [
          "V3 Workspaces"
        ]
      }
    },
...
    "/api/workspaces/{slug}": {
      "get": {
        "operationId": "WorkspacesController_getWorkspaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details"
          },
          "403": {
            "description": "Forbidden: Not a member"
          },
          "404": {
            "description": "Workspace not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get workspace details by slug",
        "tags": [
          "Workspaces"
        ]
      },
      "patch": {
        "operationId": "WorkspacesController_updateWorkspaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated successfully"
          },
          "403": {
            "description": "Forbidden: Not an owner or admin"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update workspace details",
        "tags": [
          "Workspaces"
        ]
      },
      "delete": {
        "operationId": "WorkspacesController_deleteWorkspaceBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted successfully"
          },
          "403": {
            "description": "Forbidden: Only owner can delete"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a workspace",
        "tags": [
          "Workspaces"
        ]
      }
    },
...
    "/api/workspaces/{slug}/api-tokens": {
      "get": {
        "operationId": "ApiTokensController_getApiTokens",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of API tokens"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all API tokens for a workspace",
        "tags": [
          "API Tokens"
        ]
      },
      "post": {
        "operationId": "ApiTokensController_createApiToken",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API token created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new API token for a workspace",
        "tags": [
          "API Tokens"
        ]
      }
    },
    "/api/workspaces/{slug}/api-tokens/{tokenId}": {
      "delete": {
        "operationId": "ApiTokensController_deleteApiToken",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "The workspace slug",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tokenId",
            "required": true,
            "in": "path",
            "description": "The token ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API token deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an API token",
        "tags": [
          "API Tokens"
        ]
      }
    },
...
  },
  "info": {
    "title": "Scyrme Chat API",
    "description": "Comprehensive API documentation for Scyrme Chat V2 and V3 Enterprise. This documentation describes both standard V2 integrations and high-performance, secure V3 Enterprise M2M interfaces.",
    "version": "3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "V3UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Acme Corp Updated",
            "description": "The display name of the workspace"
          },
          "icon": {
            "type": "string",
            "example": "building",
            "description": "Icon identifier or URL"
          },
          "description": {
            "type": "string",
            "example": "Updated description for Acme Corp",
            "description": "A description for the workspace."
          },
          "industry": {
            "type": "string",
            "example": "Technology",
            "description": "The industry categorization of the workspace."
          },
          "brandingConfig": {
            "type": "object",
            "description": "Custom branding configuration"
          }
        }
      },
...
      "UpdateWorkspaceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Updated Workspace Name"
          },
          "icon": {
            "type": "string",
            "example": "https://example.com/new-icon.png"
          },
          "banner": {
            "type": "string",
            "example": "https://example.com/banner.png"
          },
          "description": {
            "type": "string",
            "example": "Updated description"
          },
          "settings": {
            "type": "object"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "pro",
              "enterprise"
            ]
          },
          "isPublic": {
            "type": "boolean"
          },
          "customDomain": {
            "type": "string"
          },
          "brandingConfig": {
            "type": "object"
          },
          "industry": {
            "type": "string"
          }
        }
      },
...
      "CreateApiTokenDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "My API Token"
          },
          "permissions": {
            "type": "object",
            "properties": {
              "actions": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "read:members",
                    "write:members",
                    "read:departments",
                    "write:departments",
                    "read:teams",
                    "write:teams",
                    "read:announcements",
                    "write:announcements",
                    "read:channels",
                    "write:channels",
                    "send:messages"
                  ]
                }
              },
              "departments": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "teams": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "rateLimit": {
            "type": "number",
            "default": 1000,
            "example": 5000
          },
          "expiresAt": {
            "type": "string",
            "description": "ISO format datetime"
          }
        },
        "required": [
          "name",
          "permissions"
        ]
      },
...
    }
  }
}